requirements.txt
mcp>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0